requirements.txt
amqp==5.0.6
asgiref==3.2.10
billiard==3.6.4.0
boto==2.49.0
boto3==1.16.15
botocore==1.19.15
celery==5.1.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dateparser==0.7.6
Django==2.2
django-appconf==1.0.4
django-boto==0.3.12
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-debug-toolbar==3.2.1
django-filter==2.3.0
django-image-cropping==1.5.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-material-admin==1.7.4
django-storages==1.6.3
django-utils-six==2.0
djangorestframework==3.11.1
idna==2.10
jet-bridge-base==0.8.0
jet-django==1.0.5
jmespath==0.10.0
kombu==5.1.0
# mysqlclient==2.0.2
pilkit==2.0
Pillow==7.2.0
prompt-toolkit==2.0.9
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.1.0
pre-commit==2.17.0
